NavigationExtensions are now extensions from Fragment class
