+++ app/src/main/java/ar/com/wolox/android/cookbook/common/di/AppComponent.kt
@@ -3,6 +3,7 @@ package ar.com.wolox.android.cookbook.common.di
 import android.app.Application
 
 import ar.com.wolox.android.cookbook.CookbookApplication
+import ar.com.wolox.android.cookbook.navigation.NavigationRecipeModule
 import ar.com.wolox.wolmo.core.di.modules.ContextModule
 import ar.com.wolox.wolmo.core.di.modules.DefaultModule
 import ar.com.wolox.wolmo.core.di.scopes.ApplicationScope
@@ -16,7 +17,8 @@ import dagger.android.support.AndroidSupportInjectionModule
 @ApplicationScope
 @Component(
         dependencies = [NetworkingComponent::class],
-        modules = [AndroidSupportInjectionModule::class, DefaultModule::class, ContextModule::class, AppModule::class]
+        modules = [AndroidSupportInjectionModule::class, DefaultModule::class, ContextModule::class, AppModule::class,
+        NavigationRecipeModule::class]
 )
 interface AppComponent : AndroidInjector<CookbookApplication> {
